Move MESSAGE_DEFAULTS from config.ts to MessageDefaults.ts
