+++ universal-login-sdk/lib/MessageDefaults.ts
