+++ universal-login-sdk/test/fixtures/basicSDK.js
@@ -5,7 +5,7 @@ import {utils, ContractFactory} from 'ethers';
 import {TEST_ACCOUNT_ADDRESS} from '@universal-login/commons';
 import WalletContract from '@universal-login/contracts/build/WalletMaster.json';
 import MockToken from '@universal-login/contracts/build/MockToken';
-import MESSAGE_DEFAULTS from '../../lib/config';
+import MESSAGE_DEFAULTS from '../../lib/MessageDefaults';
 
 export default async function basicWalletService(givenProvider, wallets) {
   const [wallet, otherWallet, otherWallet2, deployer] = wallets;
